requirements.txt
# Core Django
Django>=5.0,<5.1
django-tenants>=3.6.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0

# API Framework
django-ninja>=1.1.0
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Authentication
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Async Tasks
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-redis>=5.4.0

# Environment
python-dotenv>=1.0.0

# Accounting (لاحقاً)
# django-money>=3.3.0

# PDF & Reports
reportlab>=4.0.0

# ZATCA Integration
cryptography>=41.0.0
qrcode[pil]>=7.4.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0

# Code Quality
black>=23.11.0
flake8>=6.1.0